added beter language support and submittion

+++ database/migrations/2017_10_08_203448_create_players_table.php
@@ -22,9 +22,9 @@ class CreatePlayersTable extends Migration
             $table->string('postalcode')->nullable()->default(null);
             $table->string('city')->nullable()->default(null);
             $table->ipAddress('ip');
-            $table->decimal('start',14,4)->nullable()->default(null);
-            $table->decimal('end',14,4)->nullable()->default(null);
-            $table->integer('time')->nullable()->default(null);
+            $table->decimal('start',15,4)->nullable()->default(null);
+            $table->decimal('end',15,4)->nullable()->default(null);
+            $table->decimal('time',5,1)->nullable()->default(null);
             $table->boolean('possible_dis')->default(0);
             $table->boolean('disqualified')->default(0);
             $table->string('safety_token')->default("");
